Add params keyword argument to UserAgent.urlopen, CompatRequest class, and URL class. The parameters passed in params will be urlencoded.
Remove automatic parsing and encoding of query string already set in the URL string. This make it possible to actually make a request to a URL such as 'http://some-host/get-file?path/to/file', which is currently not possible.

Remove functionality for getting and setting query parameters of URL class instances using __getitem__ and __setitem__.