Use SPDX identifier and Python packaging standard license location
As per PEP 369, many tools are expecting the license attribute to be an SPDX
license identifier - currently it is the name of the license file.

Instead, use the SPDX identifier for the MIT license, and move the license file
to the standard LICENSE.txt location as per the Python packaging documentation.